refactor: cli 관련 모듈 분리작업
